+++ tractor/_actor.py
@@ -1076,49 +1076,3 @@ class Arbiter(Actor):
 
     def unregister_actor(self, uid: Tuple[str, str]) -> None:
         self._registry.pop(uid, None)
-
-
-async def _start_actor(
-    actor: Actor,
-    main: typing.Callable[..., typing.Awaitable],
-    host: str,
-    port: int,
-    arbiter_addr: Tuple[str, int],
-    nursery: trio.Nursery = None
-) -> Any:
-    """Spawn a local actor by starting a task to execute it's main async
-    function.
-
-    Blocks if no nursery is provided, in which case it is expected the nursery
-    provider is responsible for waiting on the task to complete.
-    """
-    # assign process-local actor
-    _state._current_actor = actor
-
-    # start local channel-server and fake the portal API
-    # NOTE: this won't block since we provide the nursery
-    log.info(f"Starting local {actor} @ {host}:{port}")
-
-    async with trio.open_nursery() as nursery:
-        await nursery.start(
-            partial(
-                actor._async_main,
-                accept_addr=(host, port),
-                parent_addr=None
-            )
-        )
-        try:
-            result = await main()
-        except (Exception, trio.MultiError) as err:
-            log.exception("Actor crashed:")
-            await _debug._maybe_enter_pm(err)
-
-            raise
-        finally:
-            await actor.cancel()
-
-    # unset module state
-    _state._current_actor = None
-    log.info("Completed async main")
-
-    return result
